requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# System monitoring (optional)
psutil>=5.9.0

# Database and storage
supabase>=2.3.0
redis>=5.0.0

# Authentication and security
python-jose[cryptography]==3.3.0

# HTTP client
httpx[http2]>=0.24,<0.28
requests==2.31.0

# Vector database
pinecone-client==3.1.0

# Document processing
PyPDF2==3.0.1

# Web scraping
beautifulsoup4==4.12.2
lxml>=5.0.0

# Task scheduling
apscheduler==3.10.4

# WebSocket support
websockets>=12.0
realtime>=2.0.0

# AI and LangChain packages
openai>=1.86.0,<2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0

# Data validation and serialization
pydantic>=2.0.0

# Additional dependencies
typing-extensions>=4.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
